adds confirmation message on delete post

+++ admin/includes/view_all_posts.php
@@ -98,7 +98,7 @@ if (isset($_POST['checkBoxArray'])) {
                 echo "<td>{$post_date}</td>";
                 echo "<td><a href='../post.php?p_id=$post_id'>VIEW</a></td>";
                 echo "<td><a href='posts.php?source=edit_post&p_id=$post_id'>EDIT</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>DELETE</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \" href='posts.php?delete={$post_id}'>DELETE</a></td>";
                 echo "</tr>";
             }
             ?>
